+++ src/algorithms/constructors/mod.rs
@@ -12,6 +12,7 @@ pub mod create;
 pub mod de_bruijn;
 pub mod famous;
 pub mod full;
+pub mod full_citation;
 pub mod generalized_petersen;
 pub mod hamming;
 pub mod hexagonal_lattice;